Create exceptions.py and move InvalidUsage to it
Nice to keep exception declarations separate and out of files like
views.py
