+++ sponge/views.py
@@ -1,24 +1,9 @@
 from sponge import app
+from sponge.exceptions import InvalidUsage
 from flask import jsonify, request, abort
 import uuid
 from datetime import datetime
 from orm import db, DbRelease, DbPackage, DbResults
-
-
-class InvalidUsage(Exception):
-    status_code = 400
-
-    def __init__(self, message, status_code=None, payload=None):
-        Exception.__init__(self)
-        self.message = message
-        if status_code is not None:
-            self.status_code = status_code
-        self.payload = payload
-
-    def to_dict(self):
-        rv = dict(self.payload or ())
-        rv['message'] = self.message
-        return rv
 
 
 @app.errorhandler(InvalidUsage)
